backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
passlib==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
pydantic==2.10.5
pydantic-settings==2.7.1
email-validator==2.2.0
bcrypt==4.0.1
python-dotenv==1.0.0